Actually make sure we process all the messages in the queue before trying to receive more
This fixes the bug where sometimes tobot would read a command in chat but not send a reply until it reads one or more subsequent chats.

The cause is that the way we used to processing the queue was to take messages off the queue until the bot didn't have any more responses to send. This didn't account for the fact that the bot doesn't send responses for non-PRIVMSG messages, so if a non-PRIVMSG message added multiple queue items (say a message with several JOINs or PARTs), only the first would be processed because the bot would see no response was necessary and then assume there was nothing left in the queue.

Now, we explicitly check the size of the queue and process until there's only one queue entry (the staging area for incomplete messages).